requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
deepseek-sdk==0.1.0
defusedxml==0.7.1
distro==1.9.0
fonttools==4.56.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google_search_results==2.4.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
narwhals==1.31.0
numpy==2.2.4
openai==1.68.2
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.1
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
serpapi==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.43.2
svglib==1.5.1
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
watchdog==6.0.0
weasyprint==65.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
